+++ madym/dce_models/mdm_DCEModelDI2CXM.cxx
@@ -50,10 +50,10 @@ MDM_API std::string mdm_DCEModelDI2CXM::modelType() const
   return "mdm_DCEModelDI2CXM";
 }
 
-MDM_API void mdm_DCEModelDI2CXM::computeCtModel(int nTimes)
+MDM_API void mdm_DCEModelDI2CXM::computeCtModel(size_t nTimes)
 {
   //Reset all the model concentrations to 0
-  for (int i_t = 0; i_t < nTimes; i_t++)
+  for (size_t i_t = 0; i_t < nTimes; i_t++)
     CtModel_[i_t] = 0;
 
   for (const double& param : pkParams_)
@@ -133,7 +133,7 @@ MDM_API void mdm_DCEModelDI2CXM::computeCtModel(int nTimes)
 
   double Cp_t0 = (f_a*Ca_t[0] + f_v * Cv_t[0]);
 
-  for (int i_t = 1; i_t < nTimes; i_t++)
+  for (size_t i_t = 1; i_t < nTimes; i_t++)
   {
 
     // Get current time, and time change
